Move and refactor ide/rust:sync to tool/cargo:sync
